+++ tests/testthat/test-scale_mad.R
@@ -1,0 +1,3 @@
+test_that("scale_mad", {
+  expect_snapshot(scale_mad(mtcars$mpg))
+})